requirements.txt
requests==2.31.0
beautifulsoup4==4.12.2
selenium==4.15.2
pandas==2.1.3
numpy==1.25.2
matplotlib==3.8.2
seaborn==0.13.0
nltk==3.8.1
spacy==3.7.2
scikit-learn==1.3.2
arxiv==2.1.0
feedparser==6.0.10
newspaper3k==0.2.8
scholarly==1.7.11
PyPDF2==3.0.1
python-dotenv==1.0.0
aiohttp==3.9.1
fastapi==0.104.1
uvicorn==0.24.0
streamlit==1.28.2
plotly==5.17.0
wordcloud==1.9.2
textstat==0.7.3
# Google Gemini API
google-generativeai==0.3.2
google-ai-generativelanguage==0.4.0
# Token optimization
tiktoken==0.5.2
# Efficient text processing
python-Levenshtein==0.23.0
# Async SQLite
aiosqlite==0.19.0
# Batch processing
asyncio-throttle==1.0.2
